excel-report/requirements.txt
sqlalchemy>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
pandas>=2.0.0
